'sdfsadf'
Former-commit-id: 5810e2554ca9968e4d0e94336400f0272df0da6e

+++ Data/FuncionesGenerales.cs
@@ -2805,7 +2805,7 @@ namespace CerealNet.WSCartasDePorte
         public System.DateTime HoraArribo;
         public int brutoproc;
         public int taraproc;
-        public int Humedad;
+        public decimal Humedad;
         public int MermaHumedad;
         public int OtrasMermas;
         public int NetoFinal;
